requirements.txt
# GNR638 Assignment 2 | Requirements
# PyTorch ecosystem - Official stable cu128
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.10.0+cu128
torchvision==0.25.0+cu128
torchaudio==2.10.0+cu128

# Core ML
timm>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0

# Utilities
thop>=0.1.1.post2209072238
tqdm>=4.66.0
pyyaml>=6.0
kornia>=0.7.0
umap-learn>=0.5.5
huggingface_hub>=0.20.0
scipy>=1.11.0
